Cleaning up fake screens
